discord-client/requirements.txt
discord.py
python-dotenv
watchdog
pymongo
aiohttp
pika
pytest
asyncio
redis